# src/api/docs/openapi.yaml
openapi: 3.0.1
info:
  title: EPS API
  description: >-
    This is the API which supports the web and mobile applications for the
    Extended Prototype System (EPS).
  contact:
    email: jclements@ascend.network
  version: 0.29.0
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Logs user into the system
      description: |
        Use this endpoint to log into the EPS API. To log in, it accepts the
        accounts username or email address and password. It also accepts (but
        does not require) location data. If location data is sent, this
        information is logged as the user last location at login.
      operationId: loginUser
      requestBody:
        description: Login request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input
          content: { }
        '401':
          description: Invalid user
          content: { }
  /auth/logout:
    post:
      tags:
        - auth
      summary: Logs user out of the system
      description: >-
        This endpoint logs a user out of the system. It does this by revoking
        their JWT.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: [ ]
  /auth/request-password-reset:
    post:
      tags:
        - auth
      summary: Submit a password reset request
      description: |
        This endpoint supports submitting a password reset request. It accepts
        an email address for an account in the system and send an email to that
        address with a link to reset the account's password. If an unknown
        email address is submitted, the endpoint ignores the request and still
        returns a 200. This is so clients can't attempt to mine the system for
        valid email addresses. The email address is required and the endpoint
        will respond with a 400 if an email address is not sent.
      requestBody:
        description: Password reset request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          description: Invalid input
          content: { }
  /auth/reset-password/<string:token>:
    post:
      tags:
        - auth
      summary: Perform a password reset
      description: |
        This endpoint supports performing a password reset. It accepts the 
        password reset token as part of the URL. It also accepts a JSON body
        containing the new and confirmation passwords. Both properties are
        required and must match. Passwords must be longer than 10
        characters and shorter than 64 characters. If the password passes 
        validation and the reset password token is valid, the user's password
        will be changed.
      requestBody:
        description: Password reset object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordReset'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          description: Invalid input
          content: { }
...
  /activities:
    post:
      tags:
        - activities
      summary: Add new activity
      requestBody:
        description: |
          The endpoint accepts the following object to create new activity.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddActivityObject'
        required: true
      responses:
        '200':
          description: Success Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
    get:
      tags:
        - activities
      summary: Get all activities
      responses:
        '200':
          description: Success Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityList'
  /activities/<activity_id>:
    get:
      tags:
        - activities
      summary: Get a single activity
      parameters:
        - in: path
          name: activity_id
          schema:
            type: integer
          required: true
          description: The ID of the activity being requested.
          example: 16
      responses:
        '200':
          description: Success Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
    put:
      tags:
        - activities
      summary: Update an activity
      parameters:
        - in: path
          name: activity_id
          schema:
            type: integer
          required: true
          description: The ID of the activity that is being updated.
          example: 16
      requestBody:
        description: |
          The endpoint accepts the following object to update an activity.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivityObject'
        required: true
      responses:
        '200':
          description: Success Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
    delete:
      tags:
        - activities
      summary: Delete an activity
      parameters:
        - in: path
          name: activity_id
          schema:
            type: integer
          required: true
          description: The ID of the activity that is being deleted.
          example: 16
      responses:
        '200':
          description: Success Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualMessage'
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        login:
          type: string
          description: User's account or email address
        password:
          type: string
          description: User's password
        latitude:
          type: number
          format: float
          description: Position's latitude in decimal degrees
        longitude:
          type: number
          format: float
          description: Position's longitude in decimal degrees
        position_accuracy:
          type: number
          format: float
          description: Accuracy of the latitude and longitude properties in meters
        altitude:
          type: number
          format: float
          description: Position's altitude in meters, relative to sea level
        altitude_accuracy:
          type: number
          format: float
          description: Accuracy of the altitude expressed in meters
        heading:
          type: number
          format: float
          description: The direction in which the device is traveling
        speed:
          type: number
          format: float
          description: The velocity of the device in meters per second
      required:
        - login
        - password
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        metabase_user_id:
          type: integer
        email:
          type: string
        username:
          type: string
        role:
          # type: object
          $ref: '#/components/schemas/Role'
        name:
          type: string
        location:
          type: string
        last_seen_at:
          type: string
        is_active:
          type: boolean
        color:
          type: string
        last_login_location:
          type: object
          description: The location of the user the last time they logged in
          $ref: '#/components/schemas/Location'
        created_at:
          type: string
        updated_at:
          type: string
    ActionByUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        username:
          type: string
        name:
          type: string
        color:
          type: string
...
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        profile:
          # type: object
          $ref: '#/components/schemas/User'
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    PasswordReset:
      type: object
      properties:
        new_password:
          type: string
        confirm_password:
          type: string
      required:
        - new_password
        - confirm_password
...
    CaseHighlight:
      type: object
      properties:
        id:
          type: number
        key:
          type: string
        name:
          type: string
        definition:
          $ref: '#/components/schemas/CaseDefinitionSummary'
...
    AddCaseCustomFieldObject:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 25c0aca5-d9b1-4e4c-8b3b-7c96d954e2fe
          description: ID for the custom field that is unique with in the case's list of custom fields.
        name:
          type: string
          description: "The name or question text of the custom field."
        field_type:
          type: string
          description: The type of custom field.
          enum: [ text, textarea, select, check_box, radio_button, number, date, rank_list ]
        selections:
          type: array
          items:
            $ref: '#/components/schemas/CustomFieldSelectionObject'
          description: Array of options for the field types that have options.
        validation_rules:
          type: array
          items:
            type: string
          description: Reserved for future use.
        model_type:
          type: string
          example: Case
          description: This property must be set to "Case".
        case_definition_custom_field_id:
          type: integer
          format: int54
          example: 1
          description: The ID for the case's case definition's custom field that this custom field is an instance of.
        custom_section_id:
          type: number
          format: int64
          description: Reserved for future use.
        help_text:
          type: string
          example: This is for a single selection among many
          description: Additional help text that is displayed near the custom field.
        sort_order:
          type: integer
          format: int64
          example: 1
          description: The numerical sort order of this custom field when displayed.
        value:
          $ref: '#/components/schemas/AnyValue'
        created_at:
          type: string
          format: date-time
          example: 2020-06-9T15:44:05.387343
          description: The timestamp the record was created
        created_by_id:
          type: integer
          format: int64
          example: 1
          description: The ID of the account that created the record.
        updated_at:
          type: string
          format: date-time
          example: 2020-06-9T15:44:05.387351
          description: The timestamp the record was last updated.
        updated_by_id:
          type: integer
          format: int64
          example: 1
          description: The ID of the account that last updated the record.
...
    ContextualMessage:
      type: object
      properties:
        message:
          type: string
          description: A contextual error message.
...
    CustomField:
      type: object
      properties:
        id:
          type: number
          format: int64
          description: Unique ID for this custom field.
        name:
          type: string
          description: "The name or question text of the custom field."
        field_type:
          type: string
          description: The type of custom field.
          enum: [ text, textarea, select, check_box, radio_button, number, date, rank_list ]
        selections:
          type: array
          items:
            $ref: '#/components/schemas/CustomFieldSelectionObject'
          description: Array of options for the field types that have options.
        validation_rules:
          type: array
          items:
            type: string
          description: Reserved for future use.
        model_type:
          type: string
          description: The model this custom field is associated with.
        model_id:
          type: number
          format: int64
          description: The unique ID of the model record this custom field is associated with.
        custom_section_id:
          type: number
          format: integer
          description: Reserved for future use.
        help_text:
          type: string
          description: Additional help text that is displayed near the custom field.
        sort_order:
          type: integer
          description: The numerical sort order of this custom field when displayed.
        created_at:
          type: string
          description: The timestamp the case was created.
          format: date-time
        updated_at:
          type: string
          description: The timestamp the case was last updated.
          format: date-time
        created_by:
          $ref: '#/components/schemas/ActionByUser'
        updated_by:
          $ref: '#/components/schemas/ActionByUser'
      required:
        - id
        - name
        - field_type
        - model_type
...
    AddActivityObject:
      type: object
      properties:
        case_id:
          type: integer
          description: The ID of the case this activity is a part of.
          example: 1
        activity_definition_id:
          type: integer
          description: The ID of the activity definition that this activity is an instance of.
          example: 1
        name:
          type: string
          description: A unique name of this activity.
          example: Initial Inspection
        description:
          type: string
          description: A longer description of the activity
          example: This is the initial inspection of the construction site.
        is_complete:
          type: boolean
          description: Whether the activity is complete or note.
          example: true
        notes:
          type: array
          items:
            type: string
            example: The inspection started on time.
        latitude:
          type: number
          format: float
          description: Position's latitude in decimal degrees
        longitude:
          type: number
          format: float
          description: Position's longitude in decimal degrees
        position_accuracy:
          type: number
          format: float
          description: Accuracy of the latitude and longitude properties in meters
        altitude:
          type: number
          format: float
          description: Position's altitude in meters, relative to sea level
        altitude_accuracy:
          type: number
          format: float
          description: Accuracy of the altitude expressed in meters
        heading:
          type: number
          format: float
          description: The direction in which the device is traveling
        speed:
          type: number
          format: float
          description: The velocity of the device in meters per second
        custom_fields:
          type: array
          items:
            $ref: '#/components/schemas/AddCaseCustomFieldObject'
      required:
        - case_id
        - activity_definition_id
        - name
...
    ActivityDefinitionHighlight:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
...
    Activity:
      type: object
      properties:
        id:
          type: number
          format: int64
          example: 62
        name:
          type: string
        description:
          type: string
        is_complete:
          type: boolean
          example: true
          description: Is this activity complete or not
        completed_at:
          type: string
          example: "2020-09-30T15:37:15.197332"
          description: Timestamp this activity was set to completed.
        completed_by:
          $ref: '#/components/schemas/ActionByUser'
          description: The user that set the activity to completed.
        activity_definition:
          $ref: '#/components/schemas/ActivityDefinitionHighlight'
        case:
          $ref: '#/components/schemas/CaseHighlight'
        custom_fields:
          type: array
          items:
            $ref: '#/components/schemas/CustomField'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentObject'
        files:
          type: array
          items:
            $ref: '#/components/schemas/UploadedFile'
        surveys:
          type: array
          items:
            $ref: '#/components/schemas/SurveyHighlightObject'
        created_at:
          type: string
          example: "2020-09-30T15:37:15.197332"
        created_by:
          $ref: '#/components/schemas/ActionByUser'
        created_location:
          $ref: '#/components/schemas/Location'
        updated_at:
          type: string
          example: "2020-09-30T15:38:31.097910"
        updated_by:
          $ref: '#/components/schemas/ActionByUser'
        updated_location:
          $ref: '#/components/schemas/Location'
        history:
          type: array
          description: History of operations performed on this item. Sorted as most recent change first
          items:
            $ref: '#/components/schemas/History'
    UpdateActivityObject:
      type: object
      properties:
        name:
          type: string
          description: A unique name of this activity.
          example: Initial Inspection
        description:
          type: string
          description: A longer description of the activity
          example: This is the initial inspection of the construction site.
        is_complete:
          type: boolean
          description: Whether the activity is complete or note.
          example: true
        latitude:
          type: number
          format: float
          description: Position's latitude in decimal degrees
        longitude:
          type: number
          format: float
          description: Position's longitude in decimal degrees
        position_accuracy:
          type: number
          format: float
          description: Accuracy of the latitude and longitude properties in meters
        altitude:
          type: number
          format: float
          description: Position's altitude in meters, relative to sea level
        altitude_accuracy:
          type: number
          format: float
          description: Accuracy of the altitude expressed in meters
        heading:
          type: number
          format: float
          description: The direction in which the device is traveling
        speed:
          type: number
          format: float
          description: The velocity of the device in meters per second
    ActivityList:
      type: array
      items:
        $ref: '#/components/schemas/Activity'
  responses:
    Success:
      description: Success
    UnauthorizedError:
      description: Unauthorized
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT